Thu, Oct 19, 2023, 9:36 PM +03:00

+++ readability/readability.c
@@ -29,8 +29,8 @@ int main(void)
 
     words ++;
 
-    float L = (float) (100 * letters) / words;
-    float S = (float) (100 * sentances) / (words);
+    float L = (float) (100.0 * letters) /((float) words);
+    float S = (float) (100.0 * sentances) /((float) words);
 
     int index = (0.0588 * L) - (0.296 * S) - 15.8;
